Add __version__ attribute to Python module
